timers: Split the submodules out into their own files
